Fix lookup query
Re-add the * that allows partial word look ups, and also add quotes so
that the multi word look up works.  Without the quotes then 'tom d*'
will match a name with 'tom' in it and an address with 'd' in it.  With
the quotes, '"tom d*"' will only match names that start with the 5
characters 'tom d'.