Remove the g:notmuchconfig vim variable.
The g:notmuchconfig made it necessary to duplicate information that was most
likely already present in the environment.  This is because the NOTMUCH_CONFIG
environment variable was not respected.

The NOTMUCH_CONFIG environment variable can be used instead.  It can even be
set from within vim (the vimrc file for example), like this:

  let $NOTMUCH_CONFIG = "some/value"

See also `:help let-$` inside vim.